+++ client/src/hooks/useForm.js
@@ -3,10 +3,14 @@ import { useEffect, useState } from "react";
 function useForm( {initialValues, onSubmit}) {
   const [values, setValues] = useState(initialValues);
 
-  const handleChange = (e) => {
+  const handleInputChange = (e) => {
     const { name, value } = e.target;
     setValues({ ...values, [name]: value });
-  };
+  }
+
+  const handleChange = (stateObject) => {
+    setValues({ ...values, ...stateObject});
+  } 
 
   const handleSubmit = async (e) => {
     e.preventDefault();
@@ -17,6 +21,7 @@ function useForm( {initialValues, onSubmit}) {
   return {
     values,
     handleChange,
+    handleInputChange,
     handleSubmit,
   };
 }
